requirements.txt
flask
python-dotenv
flask-sqlalchemy
flask-migrate
mysql-connector-python